49 feedback fixing

+++ backend/feedback/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render
 from rest_framework import status, generics
 from rest_framework.response import Response
+from rest_framework.pagination import PageNumberPagination
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAdminUser, IsAuthenticated
 
 from .models import Feedback
 from .serializer import FeedbackSerializer
+from .filters import FeedbackFilter
 
 from users.models import User
 
@@ -14,13 +16,32 @@ class FeedbackView(generics.ListAPIView):
     serializer_class = FeedbackSerializer
     permission_classes = [IsAdminUser]
 
+class CommonPagination(PageNumberPagination):
+    page_size = 10
+
+    def get_paginated_response(self, data):
+        return Response({
+            'total_count': self.page.paginator.count,
+            'total_pages': self.page.paginator.num_pages,
+            'current_page': self.page.number,
+            'next': self.get_next_link(),
+            'previous': self.get_previous_link(),
+            'results': data,
+        })
+
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
 def get_feedback(request, user_id):
     user = User.objects.get(id=user_id)
-    feedback = Feedback.objects.filter(user=user)
-    serializer = FeedbackSerializer(feedback, many=True)
-    return Response(serializer.data, status=status.HTTP_200_OK)
+    feedback = Feedback.objects.filter(user=user).order_by("-updated_at")
+    filter_set = FeedbackFilter(request.GET, queryset=feedback)
+    if not filter_set.is_valid():
+        return Response(filter_set.errors, status=400)
+    
+    paginator = CommonPagination()
+    page = paginator.paginate_queryset(filter_set.qs, request)
+    serializer = FeedbackSerializer(page, many=True)
+    return paginator.get_paginated_response(serializer.data)
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
@@ -66,6 +87,8 @@ def update_feedback(request, feedback_id):
 def delete_feedback(request, feedback_id):
     try:
         feedback = Feedback.objects.get(id=feedback_id)
+        if feedback.status != "Отправлено":
+            return Response({"message": "Feedback can't be deleted"}, status=status.HTTP_400_BAD_REQUEST)
         feedback.delete()
         return Response({"message": "Feedback deleted"}, status=status.HTTP_200_OK)
     except Exception as e:
